pagination
Usually pagination is a result of the lack of available data client-side
so then we must retrieve some.  The case is different here b/c we are
only given a single set of json data.  It can be the perfect size (if
daily is selected) or overwhelmingly many (e.g. minutely is the
selector) so we paginate not b/c we lack data but to keep the size
adjusted.